Extract a generic document store not relying on vector embeddings
